move message dialog to app context
