+++ src/pages/_app.tsx
@@ -17,6 +17,10 @@ import Head from 'next/head';
 import { getAlbumSiteHost } from 'constants/pages';
 import GoToEnte from 'components/pages/sharedAlbum/GoToEnte';
 import LoadingBar from 'react-top-loading-bar';
+import MessageDialog, {
+    MessageAttributes,
+    SetDialogMessage,
+} from 'components/MessageDialog';
 
 const GlobalStyles = createGlobalStyle`
 /* ubuntu-regular - latin */
@@ -537,6 +541,8 @@ type AppContextType = {
     setRedirectURL: (url: string) => void;
     startLoading: () => void;
     finishLoading: () => void;
+    closeMessageDialog: () => void;
+    setDialogMessage: SetDialogMessage;
 };
 
 export enum FLASH_MESSAGE_TYPE {
@@ -570,6 +576,8 @@ export default function App({ Component, err }) {
     const [isAlbumsDomain, setIsAlbumsDomain] = useState(false);
     const isLoadingBarRunning = useRef(false);
     const loadingBar = useRef(null);
+    const [dialogMessage, setDialogMessage] = useState<MessageAttributes>();
+    const [messageDialogView, setMessageDialogView] = useState(false);
 
     useEffect(() => {
         if (
@@ -668,6 +676,8 @@ export default function App({ Component, err }) {
         }
     }, []);
 
+    useEffect(() => setMessageDialogView(true), [dialogMessage]);
+
     const showNavBar = (show: boolean) => setShowNavBar(show);
     const setDisappearingFlashMessage = (flashMessages: FlashMessage) => {
         setFlashMessage(flashMessages);
@@ -682,6 +692,8 @@ export default function App({ Component, err }) {
         isLoadingBarRunning.current && loadingBar.current?.complete();
         isLoadingBarRunning.current = false;
     };
+
+    const closeMessageDialog = () => setMessageDialogView(false);
 
     return (
         <>
@@ -722,6 +734,13 @@ export default function App({ Component, err }) {
             )}
             <LoadingBar color="#51cd7c" ref={loadingBar} />
 
+            <MessageDialog
+                size="lg"
+                show={messageDialogView}
+                onHide={closeMessageDialog}
+                attributes={dialogMessage}
+            />
+
             <AppContext.Provider
                 value={{
                     showNavBar,
@@ -732,6 +751,8 @@ export default function App({ Component, err }) {
                     setRedirectURL,
                     startLoading,
                     finishLoading,
+                    closeMessageDialog,
+                    setDialogMessage,
                 }}>
                 {loading ? (
                     <Container>
